rename helper.js to ratings.js
