+++ client/source/components/RR/helpers/ratings.js
